// src/app/modules/order/order.interface.ts
import { Model, Types } from 'mongoose';
import { ICow } from '../cow/cow.interface';
import { IUser } from '../user/user.interface';

export type IOrder = {
  cow: Types.ObjectId | ICow;
  buyer: Types.ObjectId | IUser;
  seller: Types.ObjectId | IUser;
  price: number;
};

export type OrderModel = Model<IOrder, Record<string, unknown>>;